# docs/openapi.yaml
...
openapi: 3.0.1
info:
  title: "jmaker-server"
  version: 1.0.0
paths:
...
  /containers/started/{name}:
    delete:
      description: ''
      responses:
        '200':
          description: ''

    parameters:
      - $ref: '#/components/parameters/containerName'
components:
  parameters:
    containerName:
      name: 'name'
      required: true
      in: 'path'
      schema:
        type: string